requirements.txt
flask==3.0.2
openai==1.12.0
python-docx==1.1.0
python-dotenv==1.0.0
markdown==3.5.2